Move all scripts in one directory.
