+++ package.json
@@ -12,7 +12,7 @@
     "serve": "docusaurus serve",
     "write-translations": "docusaurus write-translations",
     "write-heading-ids": "docusaurus write-heading-ids",
-    "check-sidebar": "node check-sidebar/index.js",
+    "check-sidebar": "node scripts/check-sidebar/index.js",
     "lint:code": "eslint . && prettier --check .",
     "lint:markdown": "markdownlint \"**/*.md\" --ignore '**/.github/vale/styles/**' --ignore '**/node_modules/**'",
     "lint": "pnpm lint:code && pnpm lint:markdown",
@@ -20,7 +20,7 @@
     "reformat:markdown": "markdownlint \"**/*.md\" --ignore node_modules --fix",
     "reformat": "pnpm reformat:code && pnpm reformat:markdown",
     "lint-staged": "lint-staged",
-    "markdown-link-check": "markdown-link-check-script/markdown-link-check.sh",
+    "markdown-link-check": "scripts/markdown-link-check-script/markdown-link-check.sh",
     "spell:error": "vale --glob='!.github/vale/styles/*' . --minAlertLevel=error",
     "spell:warn": "vale --glob='!.github/vale/styles/*' . --minAlertLevel=warning",
     "spell:all": "vale --glob='!.github/vale/styles/*' . --minAlertLevel=suggestion"
