+++ scripts/post-build-plugins/add-redirects-file.js
