+++ scripts/post-build-plugins/create-link-file.js
